--- src/app/core/store/combiner.ts
@@ -1,5 +1,0 @@
-import { ReducerMap } from './contracts';
-import { Reducer, combineReducers } from 'redux';
-
-export const combiner: <T>(map: ReducerMap<T>) => Reducer<T> = map =>
-  combineReducers(map);
