+++ src/app/core/store/index.ts
@@ -1,6 +1,4 @@
 import { StoreModule } from './store.module';
 import { MODULE_STORE_BASE_PATH } from './di-tokens';
-import { combiner } from './combiner';
-import { epad } from './epic-action-map';
 
-export { StoreModule, MODULE_STORE_BASE_PATH, combiner, epad };
+export { StoreModule, MODULE_STORE_BASE_PATH };
