+++ src/app/facade/store/login-form/login-form-duck.service.ts
@@ -5,12 +5,11 @@ import { ActionsObservable, combineEpics } from 'redux-observable';
 import { Observable } from 'rxjs/Observable';
 import { of } from 'rxjs/observable/of';
 import { delay, map, catchError, mergeMap, tap } from 'rxjs/operators';
+import { ActionFabric, AnyEpic } from '@vh/core/store/contracts';
+import { epad } from '@vh/core/store/utils';
 
 import { AppUserDuckService } from '../app-user/app-user-duck.service';
 import { ILoginFormState } from '../model/login-form';
-import { ActionFabric, AnyEpic } from '../../../core/store/contracts';
-import { epad } from '../../../core/store';
-import { UnaryFunction } from 'rxjs/interfaces';
 
 export interface ILoginFormRequestAction extends Action {
   login: string;
